"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ CSPP2-ASSIGNMENTS/Assignment-3/Solution.java
@@ -4,18 +4,18 @@ public class Solution {
 	Do not modify this main function.
 	*/
 	public static void main(String[] args) {
-		Scanner s=new Scanner(System.in);      
+		Scanner s = new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result=power(base,exponent);
+        long result=power(base,exponent);
         System.out.println(result);
 	}
-	public static int power(int base, int exponent) {
+	public static long power(int base, int exponent) {
 		if (base == 0) {
 			return 0;
 		} else if(exponent == 0) {
 			return 1;
-		} else{
+		} else {
 			return base * (power(base, exponent - 1));
 		}
 	}
